use longest match; fix $> bug; avoid SUBSEQ consing
The main contribution of this commit is to make the matching rules
more robust and intuitive. Previous to this commit, ALEXA would take
the first match in the list of matches. This is quite efficient, as it
can simply fire the rule that it sees as applicable first. However, it
gets problematic very quickly for real-world parsers.

This commit changes that behavior. Now, the longest match is searched
for among all of the lexical rules, and if multiple are applicable,
the first one is taken. This can potentially cause slight performance
degradation in pathological cases, as each of the rules must be
checked as potential candidates.

This commit also makes two additional modifications:

    - We fix a bug where $> was bound to the wrong thing.

    - Even if $@ was not used, it was still consing the matched
      string. We avoid this consing now, unless the user explicitly
      uses it.

    - Adds an EAGER option to regex expressions to allow eager
      short-circuiting of rule matching.

Thanks to Corwin de Boor (@Strikeskids) for the suggestion to
implement this.